meta: Steadied frame times, removed int narrowing.
* Now using doubles across the board. Work times appear to have decreased? :D

+++ libplip/Core/Chip8/Chip8Instance.cpp
@@ -66,7 +66,7 @@ Chip8Instance::~Chip8Instance() {
     free(m_videoOutput);
 }
 
-void Chip8Instance::Delta(const long ns) {
+void Chip8Instance::Delta(const double ns) {
     m_cycleRemaining += ns;
 
     ClearActiveBreakpoint();
