+++ libplip/Core/GameBoy/GameBoyInstance.h
@@ -26,7 +26,7 @@ namespace Plip::Core::GameBoy {
         GameBoyInstance(PlipAudio* audio, PlipInput* input, PlipVideo* video, const PlipKeyValuePairCollection &config);
         ~GameBoyInstance() override;
 
-        void Delta(long ns) override;
+        void Delta(double ns) override;
         [[nodiscard]] std::map<std::string, std::map<std::string, DebugValue>> GetDebugInfo() const override;
         std::vector<uint64_t> GetPcs() const override;
         PlipError Load(const std::string &path) override;
